requirements.txt
pytz==2023.3
python-dateutil==2.8.2
pytest==7.4.3
tabulate
typer